requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django_rest_framework
markdown
django-filter
Pillow